+++ src/main/java/utils/InputValidator.java
@@ -1,4 +1,4 @@
-package baseball;
+package utils;
 
 import constants.ExitOption;
