+++ src/test/java/baseball/InputValidatorTest.java
@@ -5,6 +5,7 @@ import java.util.HashMap;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import utils.InputValidator;
 
 
 public class InputValidatorTest {
